pyproject.toml
[project]
name = "nexus-agent"
version = "1.0.0"
description = "高并发企业级 RAG 与 Agent 编排平台"
authors = [{name = "Nexus Team"}]
requires-python = ">=3.10"
dependencies = [
    # ===== 核心 Web 框架 =====
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    
    # ===== HTTP 客户端 =====
    "httpx[http2]>=0.26.0",
    
    # ===== 数据验证 =====
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # ===== 数据库（PostgreSQL + pgvector）=====
    "sqlmodel>=0.0.14",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.2.4",
    
    # ===== 缓存（Redis）=====
    "redis>=5.0.0",
    
    # ===== 任务队列（Celery）=====
    "celery>=5.3.0",
    
    # ===== 安全（JWT + 密码哈希）=====
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    
    # ===== 工具 =====
    "tenacity>=8.2.0",
    "tiktoken>=0.5.0",
    
    # ===== Embedding 模型 =====
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    
    # ===== PDF 处理 =====
    "pdfplumber>=0.10.0",
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
    
    # ===== 异步 HTTP =====
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "async-timeout>=4.0.0",
    "anyio>=4.0.0",
    
    # ===== 监控和日志 =====
    "prometheus-client>=0.19.0",
    "structlog>=23.0.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    
    # ===== 其他 =====
    "sqlparse>=0.4.4",
    "databases>=0.8.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.26.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
]

[project.scripts]
nexus-agent = "app.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
